+++ packages/theme/src/global.ts
@@ -1,2 +1,3 @@
 export * as core from "@domphy/core";
 export * as theme from "./index.js";
+export * as palette from "./palette/index.js";