+++ packages/theme/src/index.ts
@@ -1,5 +1,6 @@
 export * from "./density.js";
 export * from "./generate.js";
+export * from "./palette/index.js";
 export * from "./size.js";
 export * from "./theme.js";
 export * from "./tone.js";